+++ Cargo.toml
@@ -49,3 +49,4 @@ tail_expr_drop_order = "allow"
 
 [workspace.lints.clippy]
 all = { level = "warn", priority = 1 }
+
