+++ opentelemetry-exporter-geneva/geneva-uploader-ffi/Cargo.toml
@@ -16,11 +16,13 @@ crate-type = ["cdylib", "staticlib", "rlib"]
 geneva-uploader = { path = "../geneva-uploader", version = "0.4.0" }
 opentelemetry-proto = { workspace = true, default-features = false, features = ["logs", "trace", "gen-tonic-messages"] }
 otap-df-pdata-views = { git = "https://github.com/open-telemetry/otel-arrow", rev = "15371dc7" }
+otap-df-pdata = { git = "https://github.com/open-telemetry/otel-arrow", rev = "15371dc7", optional = true }
 tokio = { version = "1.0", features = ["rt-multi-thread"] }
 prost = "0.14"
 
 [features]
 mock_auth = ["geneva-uploader/mock_auth"]
+otlp_bytes = ["dep:otap-df-pdata"]
 
 [lints]
 workspace = true
