+++ opentelemetry-exporter-geneva/opentelemetry-exporter-geneva/Cargo.toml
@@ -5,7 +5,7 @@ version = "0.4.0"
 edition = "2021"
 homepage = "https://github.com/open-telemetry/opentelemetry-rust-contrib/tree/main/opentelemetry-exporter-geneva/opentelemetry-exporter-geneva"
 repository = "https://github.com/open-telemetry/opentelemetry-rust-contrib/tree/main/opentelemetry-exporter-geneva/opentelemetry-exporter-geneva"
-rust-version = "1.87.0"
+rust-version = "1.88.0"
 keywords = ["opentelemetry", "geneva", "logs", "traces", "exporter"]
 license = "Apache-2.0"
 
@@ -14,8 +14,7 @@ opentelemetry_sdk = {workspace = true, default-features = false, features = ["lo
 opentelemetry-proto = {workspace = true, default-features = false, features = ["logs", "trace"]}
 geneva-uploader = { path = "../geneva-uploader", version = "0.4.0" }
 futures = "0.3"
-otap-df-pdata = { git = "https://github.com/open-telemetry/otel-arrow", rev = "15371dc7" }
-prost = "0.14"
+otap-df-pdata-views = { git = "https://github.com/open-telemetry/otel-arrow", rev = "15371dc7" }
 
 [dev-dependencies]
 opentelemetry-appender-tracing = {workspace = true}
